# Speed up active block management when player looks around