@kromka-chleba 2025-04-29
I don't like this idea, partially because blocks being processed differently is something I want to avoid because I'm aiming for a uniform simulation of the world.
Correct me if I'm wrong, but it appears to me that that the engine already loads and activates blocks that are in front of the player in a biased way.