@sfan5 2025-04-29
This feels like adding more bandaid.
What prevents us from just reducing `active_block_mgmt_interval` in general? Or is that operation slow?
Is it desirable to have active blocks not fluctuate too much?