@lhofhansl 2025-04-29
@kromka-chleba this should have no simulation difference as everything is time based anyway.

@sfan5 yes, it's a bandaid for sure. Going through all the active objects is expensive, so we can't do it too oft.