@appgurueu 2025-04-30
> Going through all the active objects is expensive, though, so we can't do it too often.

I don't think the iteration is terribly expensive. We currently do this every step for `on_step` etc. alone already.