@sfan5 2025-04-30
I maintain that we should just reduce `active_block_mgmt_interval` to be fast enough (maybe `0.8s`). And if there's a performance problem we can look at it.
It's hard to say what would be the slowest part of it. I am guessing it might be all the set operations in `ActiveBlockList::update()`.