@lhofhansl 2025-05-01
Let's not get rid of m_fast_active_block_divider from https://github.com/luanti-org/luanti/pull/12925. That one quite literally make the difference between being attached to the same object you were when you left (like a blimp or hot air balloon or whatever) and falling to you death. 0.8s is not solving that.