@sfan5 2025-05-01
> That one quite literally make the difference between being attached to the same object you were when you left (like a blimp or hot air balloon or whatever) and falling to you death. 0.8s is not solving that.

This however is somewhat specific to blocking loading and the reduced management interval matching up, see:
https://github.com/luanti-org/luanti/blob/01e439597730371f0ff2f9d89141c05c74f81597/src/serverenvironment.cpp#L952-L959
It could be solved differently.