@lhofhansl 2025-05-02
FWIW, I tried the suggestion in the comment, remembered locations removed this way, and activate them from the emerge code. That this not work, as the active object would still be stepped too late, I'm sure it's doable, but need a bit studying of the timings.