@lhofhansl 2025-05-03
I am reopening this after reviewing the discussion on https://github.com/luanti-org/luanti/pull/12925, we had a discussion there between the divider-approach and passing a flag through to the emerge thread https://github.com/luanti-org/luanti/pull/12927, and decided the divider approach was better for simplicity. If that still holds true, then also using the divider is the right approach here.

I'll re-iterate that this has not effect on the simulation speed.

- If we now favor the approach in https://github.com/luanti-org/luanti/pull/12927, I can do that for sure. In that case I'd bring back the `m_force_update_active_blocks` flag and allow `RemoteClient` to set it.
- Yet another option is allowing activation of just the active blocks/objects visible to a single client.