@lhofhansl 2025-05-04
@sfan5 Could have a look at #12927 so that I can proceed here?

We have several options:
1. Do what I propose here.
2. Reintroduce the force activation flag (before #12925) and add logic to emerge (basically #12927), and set the force flag as a player looks around.
3. Have extra logic to activate just one players active objects (not blocks). I have prototyped that and it works well. Have to think about performance, though.

I'd prefer either option (1) or (3), but am happy to revive (2) if preferred.