@sfan5 2025-05-06
As a compromise I can accept (1) + adding a TODO comment saying that `m_fast_active_block_divider` is really just a workaround and a solution similar to #12927 should be investigated.

> 4\. Separate the timings between active object in the active_block_range for objects to be enabled in the active_object_send_range_blocks

I don't understand what exactly this would mean and how it solves the problem. Can you explain?