@lhofhansl 2025-05-07
> As a compromise I can accept (1) + adding a TODO comment saying that m_fast_active_block_divider is really just a workaround and a solution similar to https://github.com/luanti-org/luanti/pull/12927 should be investigated.

That sounds good. I have since noticed that there is _still_ a delay. This is due to the fact the `RemoteClient` has the `m_nothing_to_send_pause_timer`. So ideally this should be triggered in `ServerEnvironment::step` itself.

I tried an approach where a player looking around causes its own active block to be refreshed immediately. That was more complex than I liked (need to make sure that active blocks are correctly retired, etc).

So I might think on this a bit more after all.